+++ CHANGELOG.rst
@@ -4,6 +4,7 @@ Release 0.10.0 (unreleased)
 * Support python 3.13
 * Fix too strict overlapping path check (#684)
 * Show complete URL of child manifests (#683)
+* Show remote name when using default remote (#445)
 
 Release 0.9.1 (released 2024-12-31)
 ===================================